Testing and fix: All test passed, Smashing Solved
bug found on recvInt, was expecting u16 insteed
of u 32. Client receives correct data from Server

+++ common/protocol/ReceiveProtocol.cpp
@@ -24,7 +24,7 @@ u16 ReceiveProtocol::recvShort(){
 }
 
 u32 ReceiveProtocol::recvInt(){
-    u16 num;
+    u32 num;
     if(!skt.receive(&num,sizeof(u32))){
         //throw
     }
